Delete useGetVault hook since it's unnecessary
